add client manager
